refactor admin pages (create /admin/security)
